segunda actualización jueves 07/12

+++ funcionalidades/detalle.js
@@ -1,5 +1,5 @@
 const params = new URLSearchParams(window.location.search);
-const id = parseInt(params.get('id'));
+const id = parseFloat(params.get('id'));
 
 const producto = productos.find(p => p.id === id);
